shared vue added
